separated database related operation from view controller
Separated  conference related database operation from
MainTableViewController into ConferenceDataStore class
